feat: Move form-generator to $lib
