+++ src/lib/form/example/+page.server.ts
@@ -1,4 +1,3 @@
-import type { PageServerLoad } from '../$types';
 import { testForm } from './form';
 
 export const load = (async () => {
@@ -9,6 +8,6 @@ export const load = (async () => {
     form: testForm.attributes
   }
 
-}) satisfies PageServerLoad;
+})
 
 export const actions = testForm.actions
