+++ src/lib/form/example/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { page } from '$app/stores';
-	import Form from './form.svelte';
+	import Form from '$lib/form/form.svelte';
 
 	const f = $page.data.form;
 </script>
