--- src/routes/(app)/form/+layout.ts
@@ -1,1 +1,0 @@
-export const ssr = false;
